Generate Dynamic Routes
Create dynamic routes for the statically generated pages.

As the page path depends on external data,
pages that begin with `[` and end with `]` are dynamic routes in Next.js.

Create getStaticPaths function which returns an array of possible values for id.
Create getStaticProps which fetches necessary data for the post with id.